Added files via upload
This is the Working code for the Master BT. It is to be used with the WorkingSlave program.  Together, this pair allows numbers from 0-9 to be sent from the WorkingMaster to the WorkingSlave, and the WorkingSlave responds with a message of "moving servo to...". This works each time the number is sent, the number does not have to be pressed multiple times.  Must use with CR &NL.